# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Socratic API",
    "version": "1.0.0",
    "description": "API for the Socratic platform — handles users, classes, topics, chat sessions, achievements, friends, OAuth accounts, and Google Calendar integration."
  },
  "servers": [
    {
      "url": "http://3.128.186.118:5000",
      "description": "Production EC2 server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/ingest/upload": {
      "post": {
        "summary": "Upload Document to S3 and Pinecone",
        "description": "Uploads PDF documents (syllabi, textbooks, or slides) to AWS S3, processes them with LangChain PDFLoader and RecursiveCharacterTextSplitter, and upserts them into a class-shared Pinecone namespace for Socratic AI retrieval.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "classCode"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The PDF file(s) to upload. (Supports multiple files if sent as multiple 'file' keys)"
                  },
                  "classCode": {
                    "type": "string",
                    "example": "CS 3345",
                    "description": "The unique class identifier (e.g. 'CS 3345'). This organizes documents into a shared knowledge base for all students in that class."
                  },
                  "docType": {
                    "type": "string",
                    "enum": [
                      "syllabus",
                      "textbook",
                      "slides",
                      "notes",
                      "document"
                    ],
                    "default": "document",
                    "example": "slides",
                    "description": "Categorizes the document. This helps the Socratic AI understand the context (e.g. 'According to the slides...')."
                  }
                }
              },
              "example": {
                "classCode": "CS 3345",
                "docType": "slides",
                "file": "(Binary PDF Content)"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Documents successfully ingested into S3 and vectorized in Pinecone",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "classCode": {
                      "type": "string",
                      "example": "CS 1436"
                    },
                    "namespace": {
                      "type": "string",
                      "example": "class-cs1436"
                    },
                    "docType": {
                      "type": "string",
                      "example": "syllabus"
                    },
                    "filesProcessed": {
                      "type": "integer",
                      "example": 1
                    },
                    "vectorsIngested": {
                      "type": "integer",
                      "example": 42
                    },
                    "files": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "fileName": {
                            "type": "string",
                            "example": "Fall2026_Syllabus.pdf"
                          },
                          "s3Url": {
                            "type": "string",
                            "example": "https://socratic-syllabi.s3.amazonaws.com/course-docs/CS1436/syllabus/...pdf"
                          },
                          "status": {
                            "type": "string",
                            "example": "uploaded"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "success": true,
                  "classCode": "CS 1436",
                  "namespace": "class-cs1436",
                  "docType": "syllabus",
                  "filesProcessed": 1,
                  "vectorsIngested": 42,
                  "files": [
                    {
                      "fileName": "Fall2026_Syllabus.pdf",
                      "s3Url": "https://socratic-syllabi.s3.amazonaws.com/course-docs/CS1436/syllabus/1712530000-Fall2026_Syllabus.pdf",
                      "status": "uploaded"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest/status/{classCode}": {
      "get": {
        "summary": "Get Pinecone Vector Count",
        "parameters": [
          {
            "name": "classCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "CS 1436"
          }
        ],
        "responses": {
          "200": {
            "description": "Status returned successfully"
          }
        }
      }
    },
...
    "/upload": {
      "post": {
        "summary": "Upload Syllabus PDF to S3",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "syllabusPdf",
                  "class_code"
                ],
                "properties": {
                  "syllabusPdf": {
                    "type": "string",
                    "format": "binary"
                  },
                  "class_code": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Uploaded successfully"
          },
          "400": {
            "description": "Missing file or class_code"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}